+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed Asset nodes pointing to files in subfolders being read-only and unable to be duplicated ([#459](https://github.com/verbb/navigation/issues/459)).
 - Fixed braced environment variable syntax in Custom URL nodes being misinterpreted as a node template.
 
 ## 4.0.1 - 2026-09-23